postgre

+++ ProjectFacul/ProjectFacul.Application/DTO/FatoresClimaticosDTO.cs
@@ -7,7 +7,7 @@ namespace ProjectFacul.Application.DTO
         public Guid Id { get; set; }
         public int Temperatura { get; set; }
         public int Humidade { get; set; }
-        public int IndicePluviometrico { get; set; }
+        public decimal IndicePluviometrico { get; set; }
         public DateTime Data { get; set; }
     }
 }
