+++ ProjectFacul/ProjectFacul.Domain/Entities/FatoresClimaticos.cs
@@ -8,7 +8,7 @@ namespace ProjectFacul.Domain.Entities
         public Guid Id { get; set; }
         public int Humidade { get; set; }
         public int Temperatura { get; set; }
-        public int IndicePluviometrico { get; set; }
+        public decimal IndicePluviometrico { get; set; }
         public DateTime Data { get; set; }
         public Praga Praga { get; set; }
